refact: new folder structure for linalg
