+++ xgi/linalg/__init__.py
@@ -1,2 +1,6 @@
-from . import matrix
-from .matrix import *
+from . import hodge_matrix
+from . import hypergraph_matrix
+from . import laplacian_matrix
+from .hodge_matrix import *
+from .hypergraph_matrix import *
+from .laplacian_matrix import *
